Tagging 4.1.3 -- third time's the charm?
git-svn-id: http://plugins.svn.wordpress.org/document-gallery/tags/4.1.3@1312596 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/class-thumb.php
@@ -9,7 +9,7 @@ class DG_Thumb {
 
     /**
      * TODO: Replace w/ https://codex.wordpress.org/Class_Reference/WP_Object_Cache
-     * @var array The cached copy of thumbs.
+     * @var DG_Thumb[] The cached copy of thumbs.
      */
     private static $thumbs = null;
 
@@ -189,7 +189,7 @@ class DG_Thumb {
      * @return bool Whether this instance represents a successful thumb generation.
      */
     public function isSuccess() {
-        return ! empty( $this->relative_path );
+        return ! empty( $this->relative_path ) && ! DG_Util::startsWith( $this->getPath(), DG_PATH );
     }
 
     /**
@@ -202,8 +202,13 @@ class DG_Thumb {
         // thumbs are immutable -- once created they can only be read or deleted
         if ( isset( $this->meta_id ) ) return;
 
+        // don't re-save identical thumb
+        if ( isset( self::$thumbs[$this->post_id][$this->dimensions] ) &&
+            $this == self::$thumbs[$this->post_id][$this->dimensions] ) return;
+
+
         // post_id + dimensions must be unique so purge the old entry if one exists
-        $old_thumb = self::getThumb( $this->post_id, $this->getWidth(), $this->getHeight() );
+        $old_thumb = self::getThumb( $this->post_id, $this->dimensions );
         if ( ! is_null( $old_thumb ) ) {
             $old_thumb->delete();
         }
@@ -309,7 +314,7 @@ class DG_Thumb {
      * Gets either a nested associative array mapping ID to dimension to thumb or an associative array mapping ID to thumb.
      *
      * @param $dimensions string WIDTHxHEIGHT
-     * @return array The matched thumbs.
+     * @return DG_Thumb[]|DG_Thumb[][] The matched thumbs.
      */
     public static function getThumbs($dimensions = null) {
         self::initThumbs();
@@ -330,7 +335,7 @@ class DG_Thumb {
 
     /**
      * Removes thumbs from the DB.
-     * @param $ids array|int|null Optional. The post IDs to be purged. If not given then all are purged.
+     * @param $ids int[]|int|null Optional. The post IDs to be purged. If not given then all are purged.
      * @param $blog_id null|int Optional. The blog to purge from. Defaults to active blog.
      */
     public static function purgeThumbs($ids = null, $blog_id = null) {
@@ -396,7 +401,7 @@ class DG_Thumb {
     }
 
     /**
-     * @param $thumbs array|DG_Thumb Removes files associated with given thumb(s).
+     * @param $thumbs DG_Thumb[]|DG_Thumb Removes files associated with given thumb(s).
      */
     private static function cleanupThumbFiles($thumbs) {
         if ( is_a( $thumbs, __CLASS__ ) ) {
